+++ include/extractor/command.inl
@@ -2,30 +2,24 @@
 #define EXTRACOTR_COMMAND_INL
 
 #include "core.hpp"
-#include "testfor.hpp"
 #include "extractor.hpp"
-
+#include "testfor.hpp"
 
 namespace extract
 {
 
-    template <class Modifier, std::enable_if_t<
-                std::is_same<
-                    decltype(Extractor("").test_for<Modifier>(int())),
-                    size_t>::value, 
-                    bool> = true
-                >
+    template <class Modifier, std::enable_if_t<std::is_same<decltype(Extractor("").test_for<Modifier>(int())), size_t>::value, bool> = true>
     class Command : public TestFor
     {
-    
+
     protected:
-        virtual size_t check(const Extractor* extractor, int offset) override
+        virtual size_t check(const Extractor* extractor, size_t offset) override
         {
             return Modifier::test_for(extractor, offset);
         }
-    
+
     public:
-        inline static size_t test_for(const Extractor* extractor, int offset)
+        inline static size_t test_for(const Extractor* extractor, size_t offset)
         {
             return Modifier::test_for(extractor, offset);
         }
@@ -33,6 +27,6 @@ namespace extract
         ~Command() override = default;
     };
 
-} // namespace extractor
+} // namespace extract
 
 #endif
